requirements.txt
langchain==0.0.340
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.28.2
